fixed attemp to stop openstack instance for non-cloud resources

+++ akrr/daemon.py
@@ -1221,8 +1221,9 @@ def delete_task(task_id, remove_from_scheduled_queue=True, remove_from_active_qu
             th.delete_remote_folder()
             th.delete_local_folder()
 
-            openstack_server = akrr.util.openstack.OpenStackServer(resource=th.resource)
-            openstack_server.delete()
+            if 'batch_scheduler' in th.resource and th.resource['batch_scheduler'].lower() == "openstack":
+                openstack_server = akrr.util.openstack.OpenStackServer(resource=th.resource)
+                openstack_server.delete()
         else:
             raise Exception("Task can NOT be remove safely. Unimplemented status:" + active_task['status'])
